Refactor - move getFailureType outside TransportAction
